+++ Integration/Integration.sln
@@ -11,13 +11,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "API", "API", "{F2946645-8D6
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntegrationAPI", "IntegrationApi\IntegrationAPI.csproj", "{EE270EDF-55E4-44AE-B934-8C07A2F4210E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{76D3CAAC-AAAC-4261-AF04-25DA90F395C6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntegrationClassLibTests", "IntegrationClassLibTests\IntegrationClassLibTests.csproj", "{55B62B76-51A8-4112-B724-BAD9C00FFE80}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "IntegrationUnitTests", "IntegrationUnitTests", "{42CD61FF-1EFD-438F-AC2A-2C929FBDF537}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntegrationUnitTests", "IntegrationUnitTests\IntegrationUnitTests.csproj", "{AB937E99-55A4-42AC-980D-7793CCCCE93F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntegrationUnitTests", "IntegrationUnitTests\IntegrationUnitTests.csproj", "{AB937E99-55A4-42AC-980D-7793CCCCE93F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,10 +29,6 @@ Global
 		{EE270EDF-55E4-44AE-B934-8C07A2F4210E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE270EDF-55E4-44AE-B934-8C07A2F4210E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EE270EDF-55E4-44AE-B934-8C07A2F4210E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{55B62B76-51A8-4112-B724-BAD9C00FFE80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{55B62B76-51A8-4112-B724-BAD9C00FFE80}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{55B62B76-51A8-4112-B724-BAD9C00FFE80}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{55B62B76-51A8-4112-B724-BAD9C00FFE80}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AB937E99-55A4-42AC-980D-7793CCCCE93F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AB937E99-55A4-42AC-980D-7793CCCCE93F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AB937E99-55A4-42AC-980D-7793CCCCE93F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -48,7 +40,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{A7AD5E06-A678-48E4-A92D-9926B406F9C7} = {092153ED-E218-4A6F-86DB-030C6BA9740E}
 		{EE270EDF-55E4-44AE-B934-8C07A2F4210E} = {F2946645-8D60-4CD1-AEF4-B214812022F2}
-		{55B62B76-51A8-4112-B724-BAD9C00FFE80} = {76D3CAAC-AAAC-4261-AF04-25DA90F395C6}
 		{AB937E99-55A4-42AC-980D-7793CCCCE93F} = {42CD61FF-1EFD-438F-AC2A-2C929FBDF537}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
