+++ Integration/IntegrationApi/DTO/SftpCredentialsDTO.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Policy;
+using System.Threading.Tasks;
+
+namespace IntegrationAPI.DTO
+{
+    public class SftpCredentialsDTO
+    {
+        public string Host { get; set; }
+        public string Username { get; set; }
+        public string Password { get; set; }
+    }
+}
