+++ Integration/IntegrationClassLib/Model/Reports/MedicineConsumption.cs
